move _sim_nsmb.py functions to helpers.py in test/, move other useful test functions to test/, update demo for ofnsmb to reflect changes, make symlink in demos folder to helpers.py
