+++ test/test_fitting.py
@@ -1,129 +1,10 @@
 import pytest
 import numpy as np
+
+from helpers import isclose, create_example_data, create_example_muontail
 import qetpy as qp
 from qetpy.core._fitting import _argmin_chi2, _get_pulse_direction_constraint_mask
 
-def isclose(a, b, rtol=1e-10, atol=0):
-    """
-    Function for checking if two arrays are close up to certain tolerance parameters.
-    This is a wrapper for `numpy.isclose`, where we have simply changed the default 
-    parameters.
-    
-    Parameters
-    ----------
-    a : array_like
-        Input array to compare.
-    b : array_like
-        Input array to compare.
-    rtol : float, optional
-        The relative tolerance parameter.
-    atol : float, optional
-        The absolute tolerance parameter.
-
-    Returns
-    -------
-    y : bool
-        Returns a boolean value of whether all values of `a` and `b`
-        were equal within the given tolerance.
-    
-    """
-    
-    return np.all(np.isclose(a, b, rtol=rtol, atol=atol))
-
-def create_example_data(lgcpileup=False, lgcbaseline=False):
-    """
-    Function written for creating example data when testing different
-    optimum filters.
-    
-    Parameters
-    ----------
-    lgcpileup : bool, optional
-        Flag for whether or not a second pulse should be added to the trace.
-    lgcbaseline : bool, optional
-        Flag for whether or not the trace should be shifted from zero.
-        
-    Returns
-    -------
-    signal : ndarray
-        An array of values containing the specified signal in time domain, including
-        some noise.
-    template : ndarray
-        The template for a pulse (normalized to a maximum height of 1).
-    psd_sim : ndarray
-        The two-sided power spectral density used to generate the noise for `signal`.
-    
-    """
-
-    np.random.seed(1) # need to specify the random seed for testing
-
-    fs = 625e3
-    pulse_amp = 4e-6
-    baseline_shift = 0.02e-6
-    tau_rise = 20e-6
-    tau_fall = 66e-6
-
-    f = np.fft.fftfreq(32500, d=1/fs)
-    noisesim = qp.sim.TESnoise(r0=0.03)
-
-    psd_sim = noisesim.s_iload(freqs=f) + noisesim.s_ites(freqs=f) + noisesim.s_itfn(freqs=f)
-
-    t = np.arange(len(psd_sim))/fs
-    
-    
-    pulse = np.exp(-t/tau_fall)-np.exp(-t/tau_rise)
-    pulse_shifted = np.roll(pulse, len(t)//2)
-    template = pulse_shifted/pulse_shifted.max()
-    
-    noise = qp.gen_noise(psd_sim, fs=fs, ntraces=1)[0]
-    signal = noise + np.roll(template, 100)*(pulse_amp)
-
-    if lgcpileup:
-        signal += pulse_amp * np.roll(template, 1000)
-
-    if lgcbaseline:
-        signal += baseline_shift
-
-    return signal, template, psd_sim
-
-def create_example_muontail():
-    """
-    Function written for creating an example muon tail for 
-    testing `qetpy.MuonTailFit`.
-    
-    Parameters
-    ----------
-    None
-        
-    Returns
-    -------
-    signal : ndarray
-        An array of values containing the specified signal in time domain, including
-        some noise.
-    psd_sim : ndarray
-        The two-sided power spectral density used to generate the noise for `signal`.
-    
-    """
-
-    np.random.seed(1) # need to specify the random seed for testing
-
-    fs = 625e3
-    tau_fall = 20e-3
-    pulse_amp = 0.5e-6
-
-    f = np.fft.fftfreq(32500, d=1/fs)
-    noisesim = qp.sim.TESnoise(r0=0.03)
-
-    psd_sim = noisesim.s_iload(freqs=f) + noisesim.s_ites(freqs=f) + noisesim.s_itfn(freqs=f)
-
-    t = np.arange(len(psd_sim))/fs
-    
-    pulse = np.exp(-t/tau_fall)
-    template = pulse/pulse.max()
-    
-    noise = qp.gen_noise(psd_sim, fs=fs, ntraces=1)[0]
-    signal = noise + template * pulse_amp
-
-    return signal, psd_sim
 
 def test_get_pulse_direction_constraint_mask():
     """
