+++ test/test_ofnsmb.py
@@ -1,7 +1,7 @@
-import pytest
 import numpy as np
 import qetpy as qp
-from qetpy.core._fitting import _argmin_chi2, _get_pulse_direction_constraint_mask
+
+from helpers import isclose
 
 
 def test_ofnsmb_muonsubtraction():
@@ -11,72 +11,90 @@ def test_ofnsmb_muonsubtraction():
     
     """
     
-    signal, template, psd = qp.sim._sim_nsmb.create_example_pulseplusmuontail(lgcbaseline=False)
+    signal, template, psd = qp.sim.create_example_pulseplusmuontail(lgcbaseline=False)
         
     fs = 625e3
     
     nbin = len(signal)
     
     # construct the background templates
-    backgroundtemplates, backgroundtemplatesshifts = qp.core._of_nsmb.get_slope_dc_template_nsmb(nbin)
+    backgroundtemplates, backgroundtemplatesshifts = qp.get_slope_dc_template_nsmb(nbin)
    
-    psddnu,phi,Pfs, P, sbtemplatef, sbtemplatet,iB,B,ns,nb,bitcomb,lfindex  = qp.of_nsmb_setup(template,backgroundtemplates,psd, fs)
+    (psddnu,
+     phi,
+     Pfs,
+     P,
+     sbtemplatef,
+     sbtemplatet,
+     iB,
+     B,
+     ns,
+     nb,
+     bitcomb,
+     lfindex) = qp.of_nsmb_setup(template, backgroundtemplates, psd, fs)
     
     iP = qp.of_nsmb_getiP(P)
 
     # construct allowed window for signal template
-    indwindow = np.arange(0,len(template))
+    indwindow = np.arange(len(template))
     # make indwindow dimensions 1 X (time bins)
-    indwindow = indwindow[:,None].T
+    indwindow = indwindow[:, None].T
 
     # find all indices within -lowInd and +highInd bins of background_template_shifts
     lowInd = 1
     highInd = 1
-    restrictInd = np.arange(-lowInd,highInd+1)
+    restrictInd = np.arange(-lowInd, highInd+1)
     
-    for ii in range(1,nb-1):
+    for ii in range(1, nb-1):
         # start with the second index
         restrictInd = np.concatenate((restrictInd,
-                                     np.arange(int(backgroundtemplatesshifts[ii]-lowInd),
-                                               int(backgroundtemplatesshifts[ii]+highInd+1))))
+                                     np.arange(int(backgroundtemplatesshifts[ii] - lowInd),
+                                               int(backgroundtemplatesshifts[ii] + highInd + 1))))
 
     # if values of restrictInd are negative
     # wrap them around to the end of the window
-    lgcneg = restrictInd<0
-    restrictInd[lgcneg] = len(template)+restrictInd[lgcneg]
+    lgcneg = restrictInd < 0
+    restrictInd[lgcneg] = len(template) + restrictInd[lgcneg]
 
     # make restictInd 1 X (time bins)
-    restrictInd = restrictInd[:,None].T
+    restrictInd = restrictInd[:, None].T
 
     # delete the restrictedInd from indwindow
-    indwindow = np.delete(indwindow,restrictInd)
+    indwindow = np.delete(indwindow, restrictInd)
 
     # make indwindow dimensions 1 X (time bins)
-    indwindow = indwindow[:,None].T
+    indwindow = indwindow[:, None].T
 
     indwindow_nsmb = [indwindow]
-    lgcplotnsmb=True
+    lgcplotnsmb = True
     s = signal
 
     (amps_nsmb, t0_s_nsmb, chi2_nsmb,
-    chi2_nsmb_lf,resid) = qp.of_nsmb(s,
-                                    phi, 
-                                    sbtemplatef.T, 
-                                    sbtemplatet, iP, 
-                                    psddnu.T, fs, 
-                                    indwindow_nsmb, ns, nb, 
-                                    bitcomb, lfindex, 
-                                    lgcplot=lgcplotnsmb,lgcsaveplots=False)
+    chi2_nsmb_lf, resid) = qp.of_nsmb(s,
+                                      phi, 
+                                      sbtemplatef.T, 
+                                      sbtemplatet,
+                                      iP, 
+                                      psddnu.T,
+                                      fs, 
+                                      indwindow_nsmb,
+                                      ns,
+                                      nb, 
+                                      bitcomb,
+                                      lfindex, 
+                                      lgcplot=lgcplotnsmb,
+                                      lgcsaveplots=False,
+                                     )
    
     
     priorPulseAmp = -4.07338835e-08
     priorMuonAmp = 1.13352442e-07
     priorDC = -4.96896901e-07
-    savedVals = (priorPulseAmp,  priorMuonAmp, priorDC)
+    savedVals = (priorPulseAmp, priorMuonAmp, priorDC)
     
     
     rtol = 1e-7
-    assert np.all(np.isclose(amps_nsmb, savedVals, rtol=rtol, atol=0)) 
+    assert isclose(amps_nsmb, savedVals, rtol=rtol, atol=0)
 
 
 def test_ofnsmb_ttlfitting():
@@ -88,30 +106,30 @@ def test_ofnsmb_ttlfitting():
     fs = 625e3
     ttlrate = 2e3
 
-    signal, template, psd = qp.sim._sim_nsmb.create_example_ttl_leakage_pulses(fs,ttlrate)
+    signal, template, psd = qp.sim.create_example_ttl_leakage_pulses(fs, ttlrate)
 
     nbin = len(signal)
 
     (backgroundtemplates,
     backgroundtemplateshifts,
     backgroundpolarityconstraint,
-    indwindow_nsmb) = qp.core._of_nsmb.maketemplate_ttlfit_nsmb(template, 
-                                                                  fs, 
-                                                                  ttlrate, 
-                                                                  lgcconstrainpolarity=True,
-                                                                  lgcpositivepolarity=False,
-                                                                  notch_window_size=1)
+    indwindow_nsmb) = qp.maketemplate_ttlfit_nsmb(template,
+                                                  fs,
+                                                  ttlrate,
+                                                  lgcconstrainpolarity=True,
+                                                  lgcpositivepolarity=False,
+                                                  notch_window_size=1)
 
     # add a second npdarray to the window list
     # to test the window functionality
-    indwindow_nsmb.append(np.arange(0,100))
+    indwindow_nsmb.append(np.arange(100))
 
     # concatenate signal and background template matrices and take FFT
-    sbtemplatef, sbtemplatet = qp.of_nsmb_ffttemplate(np.expand_dims(template,1), backgroundtemplates)
+    sbtemplatef, sbtemplatet = qp.of_nsmb_ffttemplate(np.expand_dims(template, 1), backgroundtemplates)
 
     (psddnu, phi, Pfs, P,
     sbtemplatef, sbtemplatet, iB,
-    B, ns, nb, bitcomb, lfindex)  = qp.of_nsmb_setup(template, backgroundtemplates, psd, fs)
+    B, ns, nb, bitcomb, lfindex) = qp.of_nsmb_setup(template, backgroundtemplates, psd, fs)
 
     sigpolarityconstraint = (-1)*np.ones(1)
     s = signal
@@ -128,21 +146,20 @@ def test_ofnsmb_ttlfitting():
      chi2_nsmb_cwindow_int,
      t0_s_nsmb_cwindow_int) = qp.of_nsmb_con(s, phi, Pfs,
                                              P, sbtemplatef.T, sbtemplatet,
-                                             psddnu.T, fs, indwindow_nsmb, ns,nb, bitcomb, lfindex,
-                                             background_templates_shifts = backgroundtemplateshifts,
-                                             bkgpolarityconstraint = backgroundpolarityconstraint,
-                                             sigpolarityconstraint = sigpolarityconstraint,
-                                             lgcplot=False,lgcsaveplots=False)
+                                             psddnu.T, fs, indwindow_nsmb, ns, nb, bitcomb, lfindex,
+                                             background_templates_shifts=backgroundtemplateshifts,
+                                             bkgpolarityconstraint=backgroundpolarityconstraint,
+                                             sigpolarityconstraint=sigpolarityconstraint,
+                                             lgcplot=False, lgcsaveplots=False)
 
 
     (ampsbonly_nsmb, chi2bonly_nsmb,
      chi2bonly_nsmb_lf) = qp.of_mb(s, phi, sbtemplatef.T, sbtemplatet,
-                                        iB, B, psddnu.T, fs, ns, nb, lfindex,
-                                        background_templates_shifts = backgroundtemplateshifts,
-                                        bkgpolarityconstraint = backgroundpolarityconstraint,
-                                        sigpolarityconstraint = sigpolarityconstraint,
-                                        lgcplot=False, lgcsaveplots=False)
-    print(ampsbonly_nsmb)
+                                   iB, B, psddnu.T, fs, ns, nb, lfindex,
+                                   background_templates_shifts=backgroundtemplateshifts,
+                                   bkgpolarityconstraint=backgroundpolarityconstraint,
+                                   sigpolarityconstraint=sigpolarityconstraint,
+                                   lgcplot=False, lgcsaveplots=False)
 
     # check the signal amplitude and the first three
     # background amplitudes
@@ -169,4 +186,4 @@ def test_ofnsmb_ttlfitting():
 
 
     rtol = 1e-7
-    assert np.all(np.isclose(newVals, savedVals, rtol=rtol, atol=0))
+    assert isclose(newVals, savedVals, rtol=rtol, atol=0)
